+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * *[v0.1.0-rc7]*
   * [FEATURE]: Add ability to delete campaigns
+  * [FEATURE]: Add ability to display instructions for setting-up
+    install tracking
+* *[/v0.1.0-rc7]*
 
 * *[v0.1.0-rc6]*
   * [FIX]: Allow users to create campaigns